back/articles/serializers.py
from rest_framework import serializers
from django.contrib.auth import get_user_model
from .models import Article

User = get_user_model()

class ArticleSerializer(serializers.ModelSerializer):
    class Meta:
        model = Article
        fields = ('id', 'user', 'img_path',)

class UserCreationSerializer(serializers.ModelSerializer):
    class Meta:
        model = User
        fields = ('id', 'username', 'password',)

